refactoring

+++ RPG/MusicalObject/AccuracyRange.cpp
@@ -8,7 +8,7 @@ AccuracyRange::AccuracyRange(size_t level)
 
 void AccuracyRange::ChangeAccuracyLevel(size_t level)
 {
-	const float accRate = (float)defaultAccLevel / (float)level;
+	const double accRate = (double)defaultAccLevel / (double)level;
 	for (int idx = 0; idx < (int)RangeName::MAX; ++idx)
 	{
 		accRange[idx] = DefaultAccRange[idx] * accRate;
